+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: allefebv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/05 18:47:30 by allefebv          #+#    #+#             */
-/*   Updated: 2018/11/15 12:20:03 by allefebv         ###   ########.fr       */
+/*   Updated: 2018/11/15 14:32:33 by allefebv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dst, const void *src, size_t len)
 {
 	unsigned char	*dst_2;
 	unsigned char	*src_2;
-	unsigned long	i;
+	size_t			i;
 
 	dst_2 = (unsigned char*)dst;
 	src_2 = (unsigned char*)src;
