+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: allefebv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/05 17:45:02 by allefebv          #+#    #+#             */
-/*   Updated: 2018/11/15 12:15:43 by allefebv         ###   ########.fr       */
+/*   Updated: 2018/11/15 14:32:49 by allefebv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t len)
 {
 	unsigned char	*s_2;
-	unsigned long	i;
+	size_t			i;
 
 	s_2 = (unsigned char*)s;
 	i = 0;
